시간별 날씨 UI 구현
- 스크롤 가능한 시간별 날씨 예보 섹션 추가
- 3시간 간격으로 2일치 날씨 데이터 표시
- mapping 단계에서 날씨 반올림

+++ domain/src/main/java/com/devhoseong/domain/model/Weather.kt
@@ -7,9 +7,9 @@ data class Weather(
 )
 
 data class CurrentWeather(
-    val temperature: Double,
-    val minTemp: Double,
-    val maxTemp: Double,
+    val temperature: Int,
+    val minTemp: Int,
+    val maxTemp: Int,
     val humidity: Int,
     val windSpeed: Double,
     val clouds: Int,
@@ -19,7 +19,7 @@ data class CurrentWeather(
 
 data class HourlyForecast(
     val timestamp: Long,
-    val temperature: Double,
+    val temperature: Int,
     val weatherIcon: String
 )
